requirements.txt
openai>=1.30.0
requests>=2.31.0
google-api-python-client>=2.126.0
google-auth>=2.29.0
